:truck: Move list-image logic and styles to own modules
