+++ packages/core/src/components/list/list-item/list-item-head/ListItemHead.tsx
@@ -8,15 +8,13 @@ import React, {
     useRef,
     useState,
 } from 'react';
-import GridImage from '../../../grid-image/GridImage';
 import Icon from '../../../icon/Icon';
 import ListItemIcon from './list-item-icon/ListItemIcon';
+import ListItemImage from './list-item-image/ListItemImage';
 import {
     StyledListItemHead,
     StyledListItemHeadBottomRightElement,
     StyledListItemHeadContent,
-    StyledListItemHeadImage,
-    StyledListItemHeadImageWrapper,
     StyledListItemHeadRightElement,
     StyledListItemHeadSubtitle,
     StyledListItemHeadSubtitleText,
@@ -56,14 +54,9 @@ const ListItemHead: FC<ListItemHeadProps> = ({
     shouldShowRoundImage,
     title,
 }) => {
-    const [hasLoadedImage, setHasLoadedImage] = useState(false);
     const [shouldShowHoverItem, setShouldShowHoverItem] = useState(false);
 
     const longPressTimeoutRef = useRef<number>();
-
-    const handleImageLoaded = useCallback(() => {
-        setHasLoadedImage(true);
-    }, []);
 
     const handleMouseEnter = useCallback(() => setShouldShowHoverItem(true), []);
 
@@ -89,32 +82,12 @@ const ListItemHead: FC<ListItemHeadProps> = ({
             return <ListItemIcon icons={icons} />;
         }
 
-        if (images && images[0] && images[1] && images[2]) {
-            const gridImages = [images[0], images[1], images[2]];
-
-            return (
-                <GridImage
-                    images={gridImages}
-                    shouldShowRoundImage={shouldShowRoundImage}
-                    size={40}
-                />
-            );
-        }
-
-        if (images && images[0]) {
-            return (
-                <StyledListItemHeadImageWrapper shouldShowRoundImage={shouldShowRoundImage}>
-                    <StyledListItemHeadImage
-                        isHidden={!hasLoadedImage}
-                        onLoad={handleImageLoaded}
-                        src={images[0]}
-                    />
-                </StyledListItemHeadImageWrapper>
-            );
+        if (images) {
+            return <ListItemImage images={images} shouldShowRoundImage={!!shouldShowRoundImage} />;
         }
 
         return undefined;
-    }, [handleImageLoaded, hasLoadedImage, icons, images, shouldShowRoundImage]);
+    }, [icons, images, shouldShowRoundImage]);
 
     return (
         <StyledListItemHead
